added mmgbsa scoring function

+++ DockTbx/autodock.py
@@ -131,17 +131,19 @@ autodock4 -p dock.dpf -l dock.dlg"""% locals()
 
     def extract_docking_results(self, file_s, input_file_r, input_file_l):
         """extract poses and scores from .dlg file"""
-        with open('dock.dlg','r') as pdbqtf:
-            with open(file_s, 'w') as sf: 
-                line = '' # initialize line
-                while 'CLUSTERING HISTOGRAM' not in line:
-                    line = pdbqtf.next()
-                    if 'Estimated Free Energy of Binding' in line:
-                        score = float(line.split()[8])
-                        print >> sf, score
 
-        subprocess.check_output('babel -ad -ipdbqt dock.dlg -omol2 lig-.mol2 -m &>/dev/null', shell=True, executable='/bin/bash')
-        self.update_output_mol2files(sample=input_file_l)
+        if os.path.exists('dock.dlg'):
+            with open('dock.dlg','r') as pdbqtf:
+                with open(file_s, 'w') as sf: 
+                    line = '' # initialize line
+                    while 'CLUSTERING HISTOGRAM' not in line:
+                        line = pdbqtf.next()
+                        if 'Estimated Free Energy of Binding' in line:
+                            score = float(line.split()[8])
+                            print >> sf, score
+
+            subprocess.check_output('babel -ad -ipdbqt dock.dlg -omol2 lig-.mol2 -m &>/dev/null', shell=True, executable='/bin/bash')
+            self.update_output_mol2files(sample=input_file_l)
 
     def extract_rescoring_results(self, filename):
         """extract scores from .dlg file"""
